Commit com algumas funções implementadas

+++ src/main/java/br/com/ticinema/cinema/DTO/ingresso/IngressoDTO.java
@@ -6,13 +6,15 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @AllArgsConstructor
 @NoArgsConstructor
 @Data
 public class IngressoDTO {
 
     private Long idingresso;
-    private Float valorTotal;
+    private BigDecimal valorTotal;
     private Integer quantidade;
     private Long clienteId;
     private ClienteDTO clienteDTO;
