+++ src/main/java/br/com/ticinema/cinema/DTO/ingresso/IngressoIngressoDTO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @AllArgsConstructor
@@ -12,7 +13,7 @@ import java.time.LocalDate;
 public class IngressoIngressoDTO {
 
     private Long idingresso;
-    private Float valorTotal;
+    private BigDecimal valorTotal;
     private Integer quantidade;
     private Long sessao;
     private LocalDate horario;
@@ -20,7 +21,7 @@ public class IngressoIngressoDTO {
     private Long cliente;
     private String nome;
 
-    public IngressoIngressoDTO(Long idingresso, Float valor, Integer quantidade, Long sessao, LocalDate horario) {
+    public IngressoIngressoDTO(Long idingresso, BigDecimal valor, Integer quantidade, Long sessao, LocalDate horario) {
         this.idingresso = idingresso;
         this.valorTotal = valor;
         this.quantidade = quantidade;
